+++ src/lib/getPoll.ts
@@ -1,7 +1,7 @@
 import { IRead } from '@rocket.chat/apps-engine/definition/accessors';
 import { RocketChatAssociationModel, RocketChatAssociationRecord } from '@rocket.chat/apps-engine/definition/metadata';
 
-import { IPoll } from './IPoll';
+import { IPoll } from '../IPoll';
 
 export async function getPoll(msgId: string, read: IRead): Promise<IPoll> {
     const association = new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, msgId);
