+++ src/lib/storeVote.ts
@@ -2,7 +2,7 @@ import { IPersistence } from '@rocket.chat/apps-engine/definition/accessors';
 import { RocketChatAssociationModel, RocketChatAssociationRecord } from '@rocket.chat/apps-engine/definition/metadata';
 import { IUser } from '@rocket.chat/apps-engine/definition/users';
 
-import { IPoll } from './IPoll';
+import { IPoll } from '../IPoll';
 
 export async function storeVote(poll: IPoll, voteIndex: number, voter: IUser, { persis }: { persis: IPersistence }) {
     const association = new RocketChatAssociationRecord(RocketChatAssociationModel.MISC, poll.msgId);
